Completed almost all firebase functionality
Sorry this taken a second, courses have been killing me :P
Added functionality to edit tasks, and projects.
Moved projects and tasks to also have a local storage in a hook in the userContext. This will allow us to stay up to date with the server without needing to keepy querying
Need to add functionality to delete tasks
Need to change how projects are named in the collection to allow for more efficient renaming. The solution to this is having a hook with document references and corresponding titles --> To check to see if the project already exists
Moved EditTask and CreateTask to their own components. Will do the same for DeleteTask
Next Steps:
- Finish implementing functionality to edit project named
- Implement functionality to delete tasks
- Implement functionality to delete projects
- Start structure for general UI layout and configuration