Made layout cards slightly better and fiddeled with layout snapping
Changed the grid layout crds to better reflect the design. But the actualy implementation did not turn out so well
Configured margins to look the initial layout design, and configured the height and width to match the design. Have not tested on smaller screens
Messed with the layout snapping when there is a height greater than 1, didn't go so well. Really need to revist this
Next Steps:
- Complete fixing the swaping of the 2 height card
- Tune the UI implementation of the cards
- Make the layout lockable, so if the user's doesn't want to accidently drag they don't have to
- Add buttons to allow configurability of cards
- Add navbar at the top (which includes next task due and holds settings and login/logout