CSS rules and dashboard loc button
Corrected CSS rules for project vie wlionk to not turn blue
Added a button to disable and enable the react grid dragable layout. This button utilizes two hooks in the Dashboard file, and I added a new class to the modules.css for the Dashboard.
I also removed the console.log from the LayoutCard
Next Steps:
- Create a way to add and remove cards
- Create a way to svae layout
- Create a way to configure the cards based on user input in edit mode
- Reexamine the the color choices for lock and unlock buttons, might be inconsistent with material UI principals
- Give button and navbar greater space