Catching up old commit and Style changes
Changed the cards to have white background instead of white based on accent
Added inter font, tried to give the title the correctw eight but form some reason it's not being called correctly.
Changed display project to use due date and time from the task like in the style guide. Might consider adding a second field to specify the accent and the light version of the color
Need to use the calendar icon for the due date on the icon
Also made some UI refinement to DashboardCardSettings and DashboardLayout