Fixed some bugs
Fixed some bugs that were preventing the edit modal to be closed
Fixed some bugs with the new way to set Projects, by updaying to new call. Got as far as delete project, so need to change the call for the rest
Bug is still presetn where wrong title is being passed to the project modal.