Integrated graph to actually pull from projects
Added ability for graph to actually pull real project data passed to it
Changed the graph to generate labels based on how many projects there are by default.
The graph then changes the label color to color set in the project. The remaining tasks is currently set to the complimentary color of the color chosen
Added a colorUtils file to help with color generation. Specifically added a function to find the complimentary color of a given a hex color.
This feature is currently set to true by default, but can be set back changing the seperateProjects flag, or passing a false value to the parameter.
Next Steps:
- Add settings to seperate projects or not in the graph
- Add settings to select specific projects to display in the graph if seperate projects is chosen
- Tighten up the label spacing of the graph to be more readable
- Change the series title to be more readible